Figure out DX input path based on actual classpath.
This synchronizes the previous work on figuring out the actual
classpath for the library container and reuse this to figure out
what goes into dex.

Also use the classpath to figure out the java resources that should
be packaged in the apk. Additionally, only add java resources by
reading the output folder instead of the source folders as this
ensure that exclusion patterns are respected.

Change-Id: Iac742f4bdefab370ec3cf7d955736e84d58ead9c