Ant build now sanitize jar files in setup task.
Previously the list of jar files was sanitized (to remove
duplicates) in the dex task, but this meant the full list
(with duplicates) was passed to proguard when building in release
mode.

This changeset move the sanitization of the jar files in the
Setup Task so that the script later only deals with a sanitized
list. The means the content of libs/*.jar for the current project
must be looked at in the task instead of later in the XML script.

Change-Id: Ib5253b80ee7c1ded004bcdad6184e0900b7a7543