AVD Manager: don't auto-close 'start emulator' when there's an error.
In the AVD Manager, don't automatically close the
'start emulator' progress dialog if the emulator
has output an error.

A few emulator messages are output to stderr which are
not errors and a few others are sent to stdout which are
not worth printing or an actually errors, so this tries
to adjust the severity accordingly and guess when the
window should be kept open to read these messages.

SDK Bug: 26564

Change-Id: I98c63b38773f395f2f0c0b6112b81b65c10c82f6