Skeleton for a SdkController app with a sevice.
This is just a basic app that has a background
service and an activity to start/stop the service.
The service has a binder interface (so that the
activities can control the service) and a listener
(so that the service can report to activities.)

The service automatically starts with the app and
there's a toggle for the user to turn it off/on.

The activity also reports errors from the service, if any.

There's a notification when the service is running.
which brings back to the control activity.

Change-Id: I67d37e3d905eb328f4d1f7d13fb118f4c7077b74