Use XML outline in the layout editor when editing text
This changeset adds an XML outline in the layout XML editor when
editing text, and the graphical outline when the graphical layout
editor tab is shown. This is complicated because it's not really
supported in Eclipse:
    https://bugs.eclipse.org/bugs/show_bug.cgi?id=1917

See the long comment in LayoutEditorDelegate.delegateGetAdapter for an
explanation of the two tricks involved:
- Using a multiplexing outline implementation to switch outlines,
  triggered by editor page switch notifications
- Using a selection listener to anticipate when the XML outline is
  attempting to react to editor selection events and return the
  non-multiplexing outline from getAdapter in that specific scenario

While I was at it I also made the outline customized to handle our
icons, similar to how the quick outline already handles it, and fixed
the quick outline to work with the new delegated editors.

Change-Id: I0b7ea8c8d8c09d2ecc1500689f9dda4b5db2b492