Add plain XML editor for typeless XML files
This changeset adds a new editor delegate, "PlainXmlEditorDelegate",
which is used for XML files that have no associated resource
type. This is used instead of the OtherXmlEditorDelegate, which is
intended for resource files.

This changeset means that if you open the lint.xml file in the root of
the project, you no longer get a form (which wasn't helpful since we
don't have descriptors for that filetype).

Change-Id: I776f7dbe2b9fc994993b26524cc2965d014496aa