SdkController: Optimize sensor display.
One issue in SdkController now is that for every single sensor
update we compute 2 string.formats, which take about forever
(around 2-4 ms each on a GN.) One is used for the UI display
so we don't need to do it when updating the sensor, instead
we can keep track of the sensor values and defer the UI string
format at display time.

This also adds code to measure timings in the sensor update
loop. It's disabled by default.

In a next CL I will replace the string.format(%g) by something
else. Using Float.toString and a StringBuilder should be
a tad more efficient since we don't need Locale-dependant
formatting (and arguably should NOT have any in fact).

Change-Id: If9e01e88c76bcd9e433df0f40c7ab28d697655e7