Turn off warnings in the command-line plugin build
The command line build of the plugin displays all warnings in all
plugins: thousands. This makes it harder to find the actual error
breaking the build, and we don't need the warnings in these logs:
it's easier to process and fix warnings within Eclipse when
developing the plugins. So, this changeset turns off warnings for
the command line build configuration.

Change-Id: Id469ad9d6f8357bb331f38562bb0aaa253849f4b