SDK Manager: remove "update sdk" action.
In v1, using the SDK Launcher.exe would issue "android update sdk"
which would pop a window ready to install. The launcher was fixed
in the sdkman v2 but the command-line flag was still supported.
Remove it since it just doesn't work the same way in the new UI.

Solves SDK issue 23294.

Change-Id: If3a30685f9a3f06888e386a918e9fd59c691c58b