SDK Manager: remove obsolete "pages" mechanism.
The SdkManager v1 UI was based on the principle of
a settings-like page selector with various pages.
For ease of transition, the v2 UI retained this
to implement the settings and about pages yet was
simply wrapping them into an swt dialog.

This cleans up the code by removing the obsolete
page registration thing and instead replaces the
pages by 2 simple dialogs.

Change-Id: I7f4f3941b5ccfd6ee636c84ebd5f06c237128a84