Add 'asset.dir' property to build.xml
The 'asset.dir' property was available in previous releases.
It was possible to set the assets directory that should be used in
the build process by defining 'asset.dir=../relative/path' in the
ant.properties file. Currently, only the "asset.absolute.dir"
property is available. While one can override this property with a
relative path, the name suggests that this shouldn't be done.

Change-Id: Ia421862c9eb736e3e1dfedaa8bdd4ff3e943941e