Prompt for package uninstall when signature differs
If a package with a different signature is already installed on the
launch device, the current behaviour is to simply abort the launch
process and print an error message to the console. The developer has
to use adb to actually do the uninstall.

This patch prompts the user if he wants to uninstall the current
package directly from within Eclipse.

Fixes #22480.

Change-Id: I8b32d19ebba511cace389696848ad9da411011bd