SDK Manager: Remember last seen source names.
One issue in the SDK Manager is that sources that
are partially installed first are loaded first.
However we used to remember the source URL but not
its name, resulting in some sources having no visible
name when loaded from installed packages.

This solves the issue by storing the known source names
locally. This expands on the local settings file that
stores whether a source is enable/disable to add more
attributes.

Change-Id: Ideca00820ccad01b4cacd997b1e43b53270fa1bc