GridLayout support work
Lots of miscellaneous fixes to the GridLayout support in ADT.  The
changes include using the GridLayout state (via reflection) to
populate the model; caching the grid model for performance, and a
bunch of fixes to the code which handles insertions and
removals. There are also some new unit tests. This is not done, but is
an improvement over the current state.

Change-Id: I4851153d3e409630c2d2024c4894d1ad1535fb47