SourceRevealer: Unify fqmn and file name based searches.
logcat provides two different ways to resolve the source corresponding
to a stacktrace: the primary method is to search for file name and
line number, and the secondary method is to search for the fully
qualified method name. Which method is used depends on a preference
setting.

This patch makes the following changes:
 - Resolution now uses both the methods - it first tries to resolve
   using the fqmn, and then improves the result using the filename/
   linenumber if that is available.
 - Since both methods are used, the preference setting can be removed.
 - If there are multiple search results, rather than always jumping
   to the first result, user will be prompted to select a result

This fixes http://code.google.com/p/android/issues/detail?id=18514

Change-Id: I1abdfe5f7e70423fb04d62f1092e1a8037537837