Allow src/doc attachement for 3rd party jars in libs/
Since those jars are added dynamically through a classpath
container, the devs cannot set the source path and the javadoc
through Eclipse UI (container don't allow editing those).

To fix this, and to make sure that both paths are picked up
not only by the current project, but also by other projects
(if the current project is a library project), the value is
set by a file sitting next to the jar file.

The file is name after the jar file, adding .properties at the
end. For instance
foo.jar -> foo.jar.properties

It can currently contain 2 properties:
src: relative or absolute path to the source folder (or archive).
doc: relative or absolute path to the javadoc.

Change-Id: I4f716374ac99479caf65453973bcb1f35ba84590