Make XML code completion work for custom views & attributes
This changeset improves the custom view handling such that XML code
completion offers any custom attributes (along with documentation
tooltips and type information). This is done by finding any
declare-styleable attributes defined in the project defining the
custom view. In particular, this will also work for the GridLayout
library project shipped as part of the android.support package.

The fix is not tied to completion; it's improving the metadata
descriptors computed for custom views, so this fix for example also
makes these custom attributes show up in the property sheet in the
layout editor.

Finally, this changeset fixes a couple of bugs in this area:
- One initialization path was not looking up custom views for unknown
  descriptors, this might be the fix for
  http://code.google.com/p/android/issues/detail?id=23020
- There was a bug in the code which looks up the namespace prefix to
  use for a given namespace URI: it would return the default Android
  prefix for some non-Android URIs.
- Small performance tweak to avoid regexp construction in a loop where
  it's not needed

Change-Id: I55dfcea6e6ea9d7c38e18a47b757678176facbd2