logcat: Remove JFace TableViewer and use SWT Table directly
This patch fixes a bunch of outstanding issues related to
scrolling in the presence of a full buffer.

Currently, the logbuffer is provided as the input model to the
TableViewer, and ViewerFilter's are used to filter the data.

This patch removes the JFace toolkit and directly works on the
SWT Table. When log messages arrive, rather than refreshing the
entire table, we can now just delete the TableItems corresponding
to the logs that were pushed out, and add new TableItems for the
incoming logs.

At steady state, this implementation performs far less work than
the previous implementation. However, during startup, this
implementation will perform more work since it does not use the
SWT.VIRTUAL bit (as all TableItems are created anyway).

Also, zebra striping has been removed to avoid appearance of flicker
when scroll lock is on.

Auto scroll lock behavior has been removed, and scroll lock button
behaves exactly like the scroll lock button in an Eclipse console.

Change-Id: Ic14487f7ad41338a581aed0ba2d85d292a584950