logcat: Validate regex patterns before creating filters.
Currently, when users create regex based filters, any error in the
regex pattern is logged. When the error is logged, context switches
to the Eclipse console view. As a result, users have to switch back
to the logcat view and fix the error.

This patch validates the text on entry and sets the text color to
be red if the pattern is incorrect. There are no context switches
as a result.

This fixes issue: http://code.google.com/p/android/issues/detail?id=22019

Change-Id: I5b571e34c1517b0a78046b56ad83c2aa8632abdb