SDK Manager: fix parsing of empty AndroidVersion codename.
SDK Bug: 29952

To reproduce the bug, install any system image with the SDK
Manager and add the line "AndroidVersion.CodeName=" in the
source.properties. When the SDK Manager loads, it will
incorrectly flag the system image as "broken" because it
can't understand the empty codename.

This fixes it by sanitizing the codename when creating
an AndroidVersion and using that class when loading the
properties instead of hand checking the codename in
various places.

Change-Id: Ie4a02739e56f576c7644b5539697c943d0082aac