Replace current complex delta visitor by pattern based ones.
The build system relies on several different delta visitor that
detect file changes to detect that needs to be rebuilt but the
implementation of those files is based on going through
folders recursively while keeping states. This is complex
and prone to errors.

The new mechanism is simply based on glob-pattern, relative
to the project root.
A set of predefined patterns stored in ChangedFileSet instance
allow quick reuse.
ChangeFileSetHelper stores those as well as allow creating
others that are based on a project config (such as the output
folder).

This first CL replace the previous delta visitors in the
PostCompilerBuilder. Replacing the ones in the PreCompilerBuilder
will require a bit more work so I prefer splitting it in another
CL.

The code that runs the pattern recognition is coming from
Ant and is stored in external/ant-glob. This is not all of Ant
but a subset of the classes for our need.
This CL does include a simple set of tests for the extracted
Ant classes.

Change-Id: I2ad1b116ffb29c9f4195d863d04f8812e87a31ca