Improve search of items into style for legacy layoutlibs.
Legacy versions uses the old API that doesn't specify the
namespace of the attribute being queried. The implementation
used the namespace of the style as the namespace of the attribute
but this make little sense.

At better implementation searches in the project's namespace
and, if the attribute is not found, then searches in the
framework namespace.

Change-Id: Ief43ecd45f108162de2b1512027d4eedf2c132db