Fix aapt handling for skipped files
Aapt occasionally emits a "Skipping <file>" message. ADT treated these
error messages as a fatal error, adding an error marker and setting
the exit status to fail. The reason for that is that aapt will
generate this error message when it encounters files in the resource
directory that start with the letter "_", which is an invalid resource
name prefix.

However, aapt *also* generates this error message when it encounters
*benign* files, such as ".git", or emacs backup files such as
"layout.xml~". Luckily, it includes the "type" of the file in the
error message.

This changeset looks at the file type, and ignores skipping file
messages for benign file types.

Change-Id: If735e19d2dca4327be06bde9c6dfdcb2c95d0196