SDK Manager: Settings to control download cache.
SDK Manager > Tools > Options now has a checkbox to
deactivate the download cache. It also indicates the
path to the cache directory, its size and a button
to clear the cache immediately.

This only affects the sdk manager files for the download
cache, it doesn't clear other cached files such as the
lint binaries.

Also restructured the settings controller a tiny bit.
There's more that could be restructured here but it's
not a priority right now.

Change-Id: I474e6155bdc041770f3f7664366d0d92bd96d9b0