ndk: Show DeviceChooserDialog only when necessary.
The ADT launcher has a cache of what device was used for a particular
launch, and reuses that device if the user has specified that the same
device should be used for future launches.

This patch moves that cache out to a separate class and reuses the
same cache for the NDK launcher as well.

Change-Id: I8aa266746e33dac82c9f16ca8877230e26c45c09