Include release names in minSdkVersion code completion
Fix the code completion in manifest files such that when completing
the minSdkVersion value, it lists not just the API numbers but the
Android version and code names for all versions. This takes advantage
of the new version name table, where the previous versions only listed
names it could look up from the installed SDKs.

This changeset also uses the same mechanism for completing
targetSdkVersion.

Change-Id: I61f06047ba4bf112da3b980c0e85325e9f9ab7c5