Fix style resource completion, and update unit tests
This changeset fixes code completion in style definitions (which did
not work when the attribute being completed was not a known Android
attribute, but was referencing one).

It also updates the various code completion unit tests, which were no
longer passing because the golden files contained the attribute types
in a different order (it now deliberately lists references after other
types, like colors). It also changes the default editors used in
completion unit tests from the legacy editor types to the new common
editor type, and closes the editors after each completion, to avoid
out-of-memory scenarios.

Change-Id: I5f947c9175524f12cb4d0a5590b0bab23a4a07dc