Allow project templates to be bundled separately
This changeset moves the various FreeMarker templates out of the ADT
plugin and over into the tools install area. The code to instantiate
templates is simplified a bit now that it no longer needs to both
handle files on disk and files read out of the .jar.

There's a new first page to the various template wizards which is
shown if the tools are not up to date (specifically, if the
$SDK/tools/templates/ directory does not exist). This page explains
that the tools have to be updated, and the Next button is disabled
until they are up to date.

This CL also contains some other tweaks suggested by Roman to the
activity-to-layout name mapping and misc code changes.

Change-Id: I3bc65f54a6b79bbeedfb917a9d34ec0d312f3526