Fix LayoutEditorMatchingStrategy to handle non-layout files
If you have both res/menu/main.xml and res/layout/main.xml, and the
menu file is open, then double clicking on the layout file would *not*
open up the layout. There was a bug in LayoutEditorMatchingStrategy
which meant that it only considered the project and base names of the
files, not whether the existing editor represented a layout file.

Change-Id: Idaf7c1b145b60510f4913896d7d0b2162744a399