Allow running on multiple devices with a single launch.
This patch provides the ability to launch an application on
all connected devices with a single launch. This applies only
to Run configurations (and not to Debug/Junit test).

UI changes:
The target tab in the launch configuration dialog provides
two options right now: Manual and Automatic. This CL adds
a third option that allows launching on all connected devices.
A drop down allows the user to further narrow down the list
to just physical devices or just emulators.

Change-Id: I721a4f41e59da24ae722b93e8ef801bc910f5442