Fix bug in manifest metadata handler
The ManifestInfo class keeps a cache of the manifest metadata, such as
the package name, minSdkVersion, etc.

However, the code which cached the minSdkVersion and targetSdkVersion
fields was incorrectly placed inside an unrelated if-block, which
meant that in some cases (in particular, manifests specifying a
default theme) these fields would not get set, and the class would
return an incorrect value of 1 for minSdkVersion, regardless of the
actual value specified.

Change-Id: I32423cad45c9f6f775b7334ee3fd9bd183497abd