Ensure we grab all of a process output.
When grabing the output of an external process, during build,
we need to make sure we get all of it since we rely on this
to parse errors to put error markers on files.

The current helper methods returned as soon as the process was
finished but this didn't mean that its output was completely read.

This could cause issue where a process would fail but no error
were read, putting no marker on any file.

The fix is twofold:
- forces the helper to blocks until all the output is read.
- put a marker on the project if the exec fails but nothing
  was output on stderr.

Change-Id: I8f0b3556ea9c4597a309627d6a14a2b0386e4e52