Synchronize window docking state between windows
The new layout editor window docking framework, used for the palette
and structure views to allow views to be reorganized and minimized,
allows the user to customize (persistently) the window configuration
in the layout editor.  However, even though the state is persisted,
this is a single global state, not a per-window state.

This changeset makes the editors automatically synchronize to the
global state when they are shown. This means that if you have two
layout editors open, and you minimize the palette in one of them, then
as soon as you switch to the second layout, the palette is minimized
in that editor tab as well.

Change-Id: I863043f654c9f48ecce25f0f2686966c68669272